klip-avatar/core_v1/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
beautifulsoup4>=4.12.0
groq
elevenlabs
moviepy
pillow
opencv-python-headless
numpy
ffmpeg-python
redis
httpx
